- merging the two versions of PhysicalLocation.  Now there are two concrete implementations, one that stores x/y as shorts, and the other as ints.  The former has the default behavior of overlfowing with read names coming from certain versions of Illumina Software, while the latter currently does not.  Leaving for future work, clean up of the various physical location like classes.

+++ src/java/picard/sam/markduplicates/util/ReadEndsForMarkDuplicatesCodec.java
@@ -66,8 +66,8 @@ public class ReadEndsForMarkDuplicatesCodec implements SortingCollection.Codec<R
 
             this.out.writeShort(read.readGroup);
             this.out.writeShort(read.tile);
-            this.out.writeShort(read.x);
-            this.out.writeShort(read.y);
+            this.out.writeShort((short)read.x);
+            this.out.writeShort((short)read.y);
             this.out.writeByte(read.orientationForOpticalDuplicates);
         } catch (final IOException ioe) {
             throw new PicardException("Exception writing ReadEnds to file.", ioe);
